pyproject.toml
[project]
name = "daily-tutor-bot"
version = "0.1.0"
description = "A Telegram bot that delivers personalized quiz questions using OpenAI"
requires-python = ">=3.11"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.4.0",
    "APScheduler>=3.10.4",
    "certifi>=2024.7.4",
    "distro>=1.9.0",
    "greenlet>=3.0.3",
    "h11>=0.14.0",
    "httpcore>=1.0.5",
    "httpx>=0.27.2",
    "idna>=3.8",
    "jiter>=0.5.0",
    "openai>=1.42.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.8.2",
    "pydantic-core>=2.20.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.4",
    "pytz>=2024.1",
    "six>=1.16.0",
    "sniffio>=1.3.1",
    "SQLAlchemy>=2.0.32",
    "tqdm>=4.66.5",
    "typing-extensions>=4.12.2",
    "tzlocal>=5.2",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "ruff>=0.7.0",
]
